requirements.txt
chainlit>=1.1.0
mcp>=1.0.0
openai>=1.30.0
requests>=2.31.0
pandas>=2.1.4
openpyxl>=3.1.2
python-dotenv>=1.0.0
python-dateutil>=2.8.2
